screen_shake (meteors, etc.) made more robust and works with aiEyes
Before:  multiple screen_shake's could queue up prior to screenshake being set on the mob setting
oldeye on the last one to queue up to be the random location from the previous screen shake
so sometimes peoples clients were stuck on a location instead of back on the mob.

Now we set screenshake before the spawn so we don't spawn up multiple screen_shakes on the same mob.

Test:  Be ghost, hammer the spawn meteors secret 10 times or so, watch as your camera returns back to your mob.

Yay.

Conflicts:
	code/modules/mob/mob_helpers.dm