ESword Changes
- All weapons now have a 'no_embed' var. If this var is set to 1 (It is
0 by default) the item will bypass checks to embed into people during
melee combat and when thrown. This is intended to be used for items such
as the esword, where it is physically impossible for it to be embedded
into someone, as opposed to something like the telebaton, where it's
simply improbably that would get embedded into anyone.

- All energy-based melee weapons have no_embed set to 1.

- All energy-based melee weapons now have a more appropriate hitsound
when active. See: blade1.ogg