SecBot Harmbaton Support
God rest my soul for coding this, but I've added support for secbots to
harmbaton people.

Currently there is no way outside of adminboos to set the var, and the
only secbot who has it set to true is the ever-spiteful Officer Buzzsky.
I might do something with emagged secbots and this in the future, but I
wanna discuss that with a few people prior to implementing it.