Xenoflora move + minor xenobio and outpost fixes
This map has the following changes:
-Xenobotany moved to xenoflora on the research outpost, one tube stop
removed, and the bits up there in the north re-arranged to fit.

- Added a few more botany-related tools to their crate and an extra box
of disks just incase the clown manages to run off with one. Also added
an emergency scythe incase the scientists get overzealous with vines.

- Moved xenobio reconstitutor console to xenoflora

-Actual pumps that let you remove atmosphere from the isolation chambers
once its put in.

Minor changes:

-There was a missing area in xenobio so I added it again

- added a reinforced window to the fauna reconstitutor containment room
so a meteor/cow/crafty drone can't  stealithly decompress the entirety
of xenobio.