Research outpost changes + minor fixes
Most major change in this is the merge of all three isolation rooms into
one mega isolation room. This design will allow you to harvest, test,
and handle artifacts that require varying arrays of atmospheric
requirements.

- It also has a silly airlock that will probably break if you misuse it,
and I'm not sure how to make it work like the old virology airlock. Also
that large syphon vent is kind of buggy in the air alarm interface. If
you can figure out how to fix either, I implore you to do so.

The rest of the changes:

- Added a window in xenobiology so a crafty asteroid can't strategically
drain the air when nobody's looking.

- Added a med scanner to virology because I'm too lazy to walk out the
front door and use a medical vendor.

- Added the manditory easter egg secret items to a vacant space in the
bulkheads that was left from the last map commit. Will you be the hero
the station doesn't deserve? Or perhaps the villain? Up to you.

- Moved that lightbulb and fire alarm in the harvesting room so the
south harvester actually functions.