Minor Hand Tele Change, Depot Change
- The Hand tele's random mode will not not select turfs inside of an
area tagged to be immune to telescience. If there is a beacon there,
however, it can still be targeted with a large teleporter and then
relayed with the hand teleporter.

- The Syndicate depot is now surrounded by asteroid on almost every
side, and a large majority of the turrets on the outside have been
removed. There  have been two additional turrets added to the interior,
and there are now two exterior turrets at the end of the cave that leads
to the airlock. This is in response to player feedback, as they thought
that the turrets covering every angle of approach was detrimental due to
people who were floating by without knowing what was ahead were getting
shot and killed. This new design is still secure, but players will be
able to more easily go around the turrets if they do not wish to attempt
to break in.