Wizard Update Part 7
(That last commit was supposed to be Part 3, not 6, but might as well
roll with it)

- Ports over the summon item spell from /tg/, it allows the wizard to
mark an item and then later recall it to his hands. If it is within
another item, that item will be recalled with the marked item inside.
- Removes RNG max charge variation on staffs, most staffs will now have
6 max charges (Rather than a 25% chance of 8, 50% of 4, and 25% of 3),
the exception being the Staff of Door Creation and Staff of Chaos, which
each have 10 charges.
- Adds summonitem to Summon Magic.
- Summon Guns/Magic now cost 1 spell slot each once again.

All ported from /tg/ with the exception of the Summon Guns/Magic cost
changes, as well as the remove of staff charge variation.