Wizard Bugfixes
- Prevents the Mutate spell from being used on z-level 2, so you can't
refund the spell while it's in use for permahulk.
- Ports in repulse spell from /tg/, it pushes away movable objects and
mobs away from the wizards. Mobs within 5 tiles will be stunned for 2
seconds, and mobs in the same tile as the wizard are stunned for 5 and
take minor damage.
- Fixes forcefields and forcewall spell by porting over
/obj/effects/forcefields.dm from /tg/. (This also fixes the mime wall,
god help us all)
- Mime forcefield nerfed to a 5 second duartion, retains 30 second
cooldown. (Yeah, let's not have those last 30 seconds.)