Reverts Incendiary Mitochondria
Instead of incendiary mitochondria being an uncounterable and
untraceable method of dealing massive mounts of burn damage to someone,
it has been reverted to what the power was originally.  It will now
ignite the user when cast. I realize this isn't very useful, but the
alternative is the power being removed altogether.