Fix for atom/add_hiddenprint()
1) add_hiddenprint was not assigning the list type to fingerprintshidden
on all atoms correctly.  This has now been resolved.

NOTES

This was what was causing the issues with the singularity restart, but I
expect this to also have many other implications.  Some bugs may develop
in live testing, but I believe this will cause far more good than harm,
as all the code expects to find a list.  (fingerprintshidden is always
declared as a list)