Ports Antag Objectives from TG
- Maroon objective: Prevent person from escaping on the shuttle, this
can include killing them. (Not in rotation)
- Escape with identity objective: Changeling objective, steal a person's
identity and ID, and escape on the shuttle with it. (Not in rotation)
- Exchange objective: 2 traitors are each given documents, one has the
red documents, one has the blue. They must meet and exchange the
documents. However, there is a 20% chance each one of them will have a
backstab objective, where he must escape with both documents. (Should be
in rotation)
- Destroy objective: Kill the station AI. (Not in rotation)
- Adds steal objective to steal the station secret documents, which are
located in the vault filing cabinet. (Should be in rotation)
- Adds special (mostly just used by ninja) objective to steal a
pinpointer. (Should be in rotation)
- Folders and filing cabinets can now store top secret documents.

All associated sprites are ported along with these features from TG.