-Changes the Plasma Solidification reaction to use Carbon, not Iron -Adds in solidification of Uranium, Gold, Iron, and Silver -Adds Osmium, Tritium, Diamond and Metallic Hydrogen as reagents (currently they do nothing) and solidification reactions for each -All solidification reactions require 5 parts carbon, 5 parts frost oil and 20 parts of the mineral to be solidified -Iron changed to Carbon in solidification to allow solidification of Iron
This is mostly setup for future reactions, and quality of life to make ground minerals a two-way street, instead of plasma being special.
And yes, I am aware that diamonds are made of carbon, but in this case the grinder is making them into a powder rather than the carbon reagent (and because solidifying carbon into diamonds would be OP).