Copy-Paste Port of /tg/station's Shadowlings
Ports tgstation/-tg-station#8157

This won't compile, needs to be bugfixed and tested. The only reason I
am PRing  this right now is so that we don't have multiple people
working on the same thing.

@ZomgPonies This will require PROBABILITY SHADOWLING # to be added to
game_options.txt, and will likely require SQL database additions for the
BE_SHADOWLING antag preference.