Adds glass tables, glass table frames
This commit adds glass tables (credit to WalterJe of -tg-'s fourms for
the sprites) to the game. These can be constructed with two metal
sheets, and two glass sheets. Flipping the table causes it to shatter
and deconstruct itself. You can deconstruct the table at any stage
with a wrench.