Fixes an oversight with NO_BREATHE and corpses
This fixes:
 - Issue #778 on https://github.com/ParadiseSS13/Paradise

There was no sanity checking to test if mobs in range of a corpse have
the NO_BREATHE flag. This commit adds such a check. It does not make
sense for species like machine people to vomit. It may make sense for
Slime People and Dionae, but, after all, it's sentient slime and a tree
thing. Are they are even capable of regurgitating their food?