Selectable default languages
This commit allows any mob which can speak to choose one of their known
languages as 'default', which makes anything that they say with no
prefix (EG; say 'lol'), use that language.

This commit also changes the 'galactic common' language to use :9
instead of :0, as :0 does not work due to BYOND.. strangeness.

Implements two new IC-tab verbs and updates the check-known-languages
panel to also integrate with this system.