[WIP] Medical Doctor Hypospray
This adds a hypospray intended for medical doctors, and contains Saline-Glucose Solution on spawn. 
This is here to help doctors, and make it a bit easier for them, as it's already rough, and that little extra will help, and the list of drugs is a rough draft, some of the meds you can od on will probably be modified pending feedback.
Currently the allowed list does not work, i suspect due to it being a datum, and allowed is for objects.
Any help to get the restricted list of reagents to work would be greatly appreciated
Other changed to come include adding it to spawn in backpacks/pockets of medical doctors and paramedic, make sure it does not count as a steal objective, and possible give the CMO hypospray a minor makeover.