Adds Barsign Picking
This commit allows the bartender, or anyone with access_bar, to change the
barsign's image by hitting it with an ID. This commit also adds a variable
to disable such behavior for custom maps. Finally, this commit adds two
new icons for the barsign, named simply 'on' and 'off', which look like
the base sprite for most of the other barsigns, but blank.