Add afterbuckle proc to beds, refactor comfychairs
This commit ports Baystation12/Baystation12#8527, as well as their
afterbuckle system. After a mob is buckled/unbuckled, afterbuckle will be
called on the object, which does nothing by default.

Comfy Chairs have been refactored to use a grayscale sprite, and also to
use overlays for the arms. This means that new colors are just a RGB
definition.