LINDA patch- Airlock atmospass fix, Runtime fix
This commit fixes the bug encountered wherein airlocks were not checked
before gas was allowed to move onto their tile- but only if opened and
closed. The problem was that they did not issue the correct atmos updating
command when they closed, but they did when they opened, and when they are
created.

The runtime was seemingly random, but the small sanity check will prevent
it from happening again.