Resist update to -tg- standards
This commit updates the /mob/living resist verb to -tg- standards. It is
no longer a giant mess of 1 single verb, it has been split into a central
verb with multiple sub-procs.

This is not a port of -tg-'s optimized resist code, it simply makes the
existing code a bit less terrible, and organized in sub procs like -tg-