Autogibber + Minor patches
This replaces the old autogibber with a spawnable version that is actually
an automatic gibber. This also fixes a few minor issues I did not notice
before with the gibber code. Namely, overlays getting messed up.

The gibber will now spit out things inside of it 1 by 1 instead of all at
once.