Fix: Event manager list index out of bounds
This commit prevents a runtime of "list index out of bounds" with events
by adding a simple sanity checking system, so that the list checked always
has a valid check, regardless of if the event has a valid variable. NOTE:
May result in unintended behaviour if an event tries to complete with an
invalid severity level. It will always be set to moderate, as that is a
safe assumption for the code to make.