Brooms, Stick-horses, and Flying Wizards!
A port of https://github.com/d3athrow/vgstation13/pull/4377

![image](http://i.gyazo.com/dd415093be904dc9e5d40570598eddb9.gif)

Allows brooms to be wielded with 2-hands. Doing so places the broom
between your legs so you look like you are riding it.

You can also put a horsehead mask onto the broom to create a stick-horse
to ride. This can also be wielded, but will ALWAYS appear between your
legs.

And for the more magically inclined, wielding a broom or stick-horse (2
hands!) will cause you to ride it like a witch, giving you the ability
to bypass hazards such as tables and banana peels.