Gives blob a fighting chance
From what I have seen, most blobs die due to there being too few at
start, thus i propose this to made the minimal ammount of cores 2, and
decrease the needed players per core, it is a tentative change at best
as this can easily tip into the favor of the blob, but due to the fact
that a fairly low amount of players are ready at round start, that
should hopefully not be an issue.

Also, updated the Quarantine law the ai get with the one on the station.