Fluff item restoration - Flasks, Lighters
This commit readds every fluff lighter to a new lighters.dm file, as well
as moving every lighter into it. There are enough to justify seperating it
from cigs_lighters.dm. Fluff lighter paths have not been touched for
compatibility to the database and map.

This commit also readds the flasks, with the fluff paths removed.