Biogenerator Biomass Value Fix
Fixes the Biogenerator only giving one point of biomass per product
inserted.

The biogenerator previously calculated the points based on the amount of
the nutriment reagent in the products inserted.

Now the biogenerator will calculate the points based on the combined
amounts of the nutriment and plant-matter reagents in the product.

Overall, this just restores the values back to their pre-Nutrient
Changes levels, as nutriment in plant products was replaced with
plant-matter for nearly every plant that had it.