Botany Belt, Biogenerator Additions
Adds in the new Botany Belt, a belt designed for handling all sorts of
botany-related tools.

Adds the Botany Belt, Security Belt, Medical Belt, Janitor Belt, and
Bandolier to the Biogenerator's list of products.

Re-arranges the biogenerator's product listing to have a new "Belts"
section, dedicated to (you guessed it) belts.
- Moves the Utility (tool) into the Belts section, previously was in the
Leather section
- The new belt listings are also in this section.

All belts in the biogenerator have the same base cost of 300 points.
Increasing the efficiency of the machine will lower the costs
appropriately.

Adjusts some code related to the security belt so they won't magically
have a seclite in them when created via biogenerator.
- The belts in the sec lockers will still have their light
- The ERT belt no longer needs to clear contents during its New() proc
to remove the seclite.