New admin sound verb; Play a sound originating at every radio.
Note, verb may not work as one would expect it to. It just uses playsound
originating at each intercomm, therefore anyone that is out of range of
the intercomm will not start hearing it when they get in range, so walking
down the halls isn't going to keep playing the sound as one would expect.

However, the ability to play a sound which, when people are in the right
position, will appear to originate from the intercomms, could be useful
for an event.