Feedback Fixes
Changes instances of \red and \blue to use spans as per suggestions in
glass.dm and fishtank.dm

Fixed a mistake preventing the building of one direction windows and
fishbowls (whoops!)

Fixes grammar concerns in multiple messages

Holy water is now (magically?) worth 110% of it's reagent amount when
filling aquariums.
- This is a rather small boost to having the chaplain bless the water
you are going to dump into a filtration system anyways.

Moves the fishwater ingestion method out of the on_mob_life proc to
prevent spam, now will display when initially ingested

Updates the CritterCare vendor's sprite to include the pug and fox, as
well as adjusting the "static" between displays to give it a slight
variety.