Progress and Bugfixes
qdeled alot of things
Added in the Spell Exsanguinate
Spell Quirk: Exanguinate and whisper only work properly when phased in.
I am assuming this is due to the fact when you phase out you transfer to
a holder dummy. Will need to think on this when my brain is more active.