Now with more colored blood
Changed the GibEat() proc to use the mobs blood color. (defualts to
human gibs if we eat Ian)
Added in (or tried to) so gib and innard throwing.
TRIED TO make sure we cannot sink into robot debris.

Reworked OBjectives slightly.

Should I consider not letting this thing EAT CYBORGS?!