Fixes generic wardrobe contents
This fixes the wardrobe base type creating items, which was fine when
wardrobes didn't call their parent's New proc, but was causing literally
every wardrobe to contain some junk now that they do.

The base wardrobe type's items have been moved to a new "generic"
wardrobe, which is now used in the derelicts and an unused away mission,
replacing mapped instances of the base type.