Blob can attack living things
Fully equipped Kidan take 4,4 brute.
Fully equipped Human take 7 brute
Fully equipped IPC take 9.75 brute
This is security jumpsuit + riot suit and helmet
Anyone else takes 40-50 damage armor depending, currently there appears
to be an issue with the damage as instead of dealing 20-30, so that is a
bug that needs to be fixed.