Various Tweaks
Moved to mousedrop system. Meaning it will need to wait for other things
to go to that system but i am not in a rush,
Put the head sprites back into head DMI.
However i am having an issue with the sprite appearing on the holders
head now.
For somereason i have to tell it to put it in the active hand or it no
work.

Took the Viro bits out of the bioscan attempted to adjust the window
sizeing.

I probaly missed something.