Changeling sting nerfs/changes/tweaks
- Prevents transform sting from working on species with NO_BLOOD.
- Prevents DNA extraction sting from working on species with NO_BLOOD.
- Prevents Plasma Men from being changelings, largely for the same
reason Slime People can't be changelings, horribly mutilating people
with transform sting.
(Alternative way to do this would be for a species check in the DNA
sting's Click() proc, which is horrendous because DNA does not contain a
species datum, just a string with the species' name.)
- Ported some nerfs from tgstation/-tg-station#10482, transform stinging
somebody is no longer silent. It won't point you out specifically, but
it will create a visible message for everyone in view range, and will
make the person jittery. Also forces a cooldown on the use of all
changeling powers.
- Re-enables cryo sting. This should now be working since mob
temperatures were fixed a while ago.