Added a new admin panel, called the stationbuse panel.
This new panel is specifically intended for station-wide events, such as
all doors being bolted or opened, or a full fire-drill, those such things.

It is restricted to requiring +EVENT or +SERVER to even open.

Bit feature lacking at the moment, but relatively easy to add new buttons
Current buttons:
 - Bolt every door on-station
 - Unbolt every door on-station, disable autoclose, and open them.

 - Trigger every firealarm on-station.