Adds subtypesof helper proc
It returns a list containing all subtypes of the object path given to
it, but not the object path itself. This only exists because
typesof(/obj/item/thing) - /obj/item/thing is used quite often.

Also adds a "/" to the start of a lot of the list helper procs that did
not start with one.