Ring-sound system for PDA's; Ringtones make sounds
This commit adds a system which allows a pre-defined custom sound for the
noise a PDA makes when recieving a message, based on what the 'ringtone'
is set to. Ringtones without an entry will make the default beeping
noise.

Included special noises;
 Explosion for 'boom'
 Slipping noise for 'slip'
 Bikehorn for 'honk'
 Vox shriek for 'SKREE' (Capitals matter)
 Shortened holy ambience music for 'holy'
 Xenomorph hiss for 'xeno'

Also included; New folder, /sound/items/PDA/, specifically for the
ringtones.
Includes:
 __credits.dm
   Used for giving credit to the source of a sound file in said folder,
   for sounds licensed under such a license that requires credit.
   Not included in the compilation.

 ambicha4-short.ogg
   Shortened version of sound/ambience/ambicha4.ogg; Used for 'holy'
   ringtone.