New proc to replace world.log <<, log_to_dd()
This commit changes every 'world.log <<' message with a loggable proc-
log_to_dd().

This is adjustable in the config; If LOG_WORLD_OUTPUT is present, all
things sent to world.log will show up in the standard /data/logs/ logs.

These logs will contain the following (in order):
 Timestamp
 "DD_OUTPUT:"
 The message.

The config option for this, by default, is turned off.