Fixes #1776, Vox Welding Helmet now Visible
Title explains the whole PR, for whatever reason it was looking for the sprite in the eyes section instead of the head section.

However, this turned up another issue. The mask's UP state is the exact same sprite as humans'-- which is just odd on Vox. I'll look further into that