Vox Coloured Uniforms
Adds a Vox sprite for every single coloured uniform in color.dm and
implements it.

Had this PR waiting 'cause the stupid way I do things only allows me to
do one at a time, haha.

Anyway, you're going to have to use your imagination for this one-- as
it would be an insane amount of work to photoshoop every single sprite
onto a Vox and make a rainbow out of the lot of them.

I really wish my Dream Daemon wasn't broken.