Moves Species_fit and Sprite_sheets to /obj/item/clothing/under/color, Removes Jerseys
Greatly reduces code as per Fox' suggestion. I outright snipped out the red and blue jerseys at the bottom because in the human icons/mob/uniform.dmi there aren't even sprites for it, and I've never ever heard of it being used. The blackf suit appears to use the same sprite as the black suit, so I figured that'd be alright too.