Autohiss
This commit adds a system (fully client-toggled) to allow unathi and
tajaran to "auto-hiss". This means that, for unathi, "S" will be
automatically extended to "Ss", "Sss", or "Ssss". Which one it changes is
based on a random pick().

There are three modes, off, basic, full.
Off, is obviously, off. It does nothing. This is the default setting.
Basic, for unathi, enables the S extensions, and for tajaran, enables the
"R" extending.
Full, for unathi, enables the S extensions, alongside replacing "X" with
"Ks". It does not currently do anything extra for Tajaran.

Sinta'Unathi is unaffected for Unathi, Siik'Tajr is unaffected for
Tajaran.

On baystation, NO_STUTTER languages would also be unaffected (the noise
language). However, we do not have NO_STUTTER languages. So it hilariously
does affect the noise language. "Zartulnal Hharar dancess."