Make EXPERI-Mentor SetTypeReactions() way more efficient
Thanks, Redbook.

Basically, this used to spawn a shitload of instances and then qdel them;
Just so it could check their icon_state and see if it's not null. However,
spawning so many instances is quite intensive. An undocumented feature
(documented in the redbook, a list of undocumented features); You can
directly access a compile-time variable with just it's path by using
initial().