Adds More Vox Hair (Head+Facial)
A straight-up port from VG. Saw that someone on Bay went for it and this
was the first I've seen of VG's different Vox hairstyles so I figured
I'd make a pull here too.

I know some of these look weird, so if there's any that you absolutely
don't want in, just specify them and I'll axe it.