Hopefully fix Spacepod bullshit.
I cleaned up part of my giant mess, and tried to ensure that the
references will never be incorrect, however THAT PROBABLY WON'T FUCKING
WORK, so I also added a proc spacepods call every time they want to
interface with the occupant vars, which checks that the occupants are
actually valid, and nulls them/moves them as appropriate.
If this happens, it will make a debug log message and yell about it.