Claw Games!
Adds claw games, the absolute most-infuriating arcade game known to
mankind.
- Uses new fancy HTML5 / Javascript code stuff
- Is buildable!
- Science can produce the board with programming 2
- Requires a matter bin, manipulator, some cables, and a sheet of glass
- Costs money!
- Accepts cash or ID swipes, 15 credits per play token
- Dispenses prizeballs!
- Open them for a random plushie, carp plushie, or action figure!
- Two sets of sprites, thanks to FoS and Stephanov

Notes for the coders / future:
This provides a basic example for introducing new HTML5 / Javascript
arcade games, as well as a base type machine to use for such games.
While this may seem silly, this proves that we are not doom to have to
tolerate only the battle game and Orion Trail forever, and can even make
our own games for the arcade.