Naughty List Revenge
Santa has refused to deliver presents to the NSS Cyberiad for an unknown
reason.

Meanwhile, a distress beacon is received from the North Pole, on an
unencrypted Syndicate frequency. Apparently, Christmas Eve was anything
but a silent night.

A rogue operative seems to have infiltrated the North Pole and attempted
to coerce Santa into removing him from the Naughty List, but his plans
may have backfired. Santa has finally snapped, and must be put down in
order for the crew to receive their gifts this year.

The gateway has been reconfigured to lock onto the general location of
the distress signal, and will not lock on to any other destinations
until after Christmas has passed. Be warned, however, that the frozen
north is far from hospitable, and a variety of dangers await you beyond
the safety of the blast doors.

This mission will offer some new challenges, and is designed to be
undertaken by a squad of players. Will your team be able to put Father
Christmas to rest and rescue the presents, or will you meet your demise
in the freezing tundra?